util-linux: re-enable libmount and binaries on Microblaze
In commit 442aa88f95d6c4a921aa3d4de91f54d50bd0cd35 ("util-linux: bump
version and revamp options"), Gustavo disabled util-linux libmount and
binaries on microblaze, as it was not building properly.

However, as mentionned in the comment, these options were disabled on
Microblaze due to "libc lacks UTIME_NOW & UTIME_COMMIT for
libmount". This was true specifically for the microblaze external
toolchain that we were using at the time. But we are no longer using
this external toolchain (which proved to be broken in many ways), and
have microblaze support in our internal backend.

I have verified that with our internal toolchain, util-linux with
libmount and the binaries enabled builds fine.

Those options are not selected by anything else in Buildroot, so
there's no other package impacted by this dependency change.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>