uclibc: enable UCLIBC_HAS_STRING_GENERIC_OPT in all configs
Description in uClibc for this optoin is:
--->---
bool "Use faster (but larger) generic string functions"
	default y
	help
	  Answer Y to use the (tweaked) glibc generic string functions.

	  In general, they are faster (but 3-5K larger) than the base
	  uClibc string functions which are optimized solely for size.

	  Many people will answer Y.
--->---

Additional 3-5K of size is not that important if resulting libuClibc.so is
of size about 300k.

But benefits in terms of performance could be pretty significant.

For example on ARC in LMbench we see more than 3 times bump in some tests
--->---
libc bcopy unaligned
...
8.39 - 6.76  (UCLIBC_HAS_STRING_GENERIC_OPT=no)
8.39 - 23.86 (UCLIBC_HAS_STRING_GENERIC_OPT=yes)
--->---

Also since the option in question is default in uClibc I don't expect if
enabled back (it was silently disabled in http://git.buildroot.net/buildroot/commit/?id=cfbf8abc33d86a0cf5c1bb3e0817a22009b7f301 on introduction of NPTL in uClibc
by Khem Raj) it to introduce problems.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Anton Kolesov <anton.kolesov@synopsys.com>
Cc: Peter Korsgaard <peter@korsgaard.com>
Cc: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>