configs: add a DT-enabled Raspberry Pi defconfig
This is an example of a Device-Tree-enabled Raspberry Pi defconfig.

We have to use a 3.18-based kernel for that, but there are a few
limitations:

  - we can not use the minimalist RPi defconfig bundled with the kernel,
    namely bcmrpi_quick_defconfig, because it is not DT-enabled, and
    sets CONFIG_ARM_PATCH_PHYS_VIRT to 'n', which prompts a value for
    CONFIG_PHYS_OFFSET, as it as no default for the bcm familly;

  - most importantly, the rpi-3.18.y branch is constantly rebased, so
    there is no guarantee that the sha1 I use today will still be usable
    in the long term. Using the name of hte branch is not better either.

So, we bundle our own DT-enabled linux defconfig that is based on
bcmrpi_quick_defconfig, with just CONFIG_ARM_PATCH_PHYS_VIRT set and the
Device TRee enabled.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>