read-edid: make available for !x86
It is only the get-edid tool that is x86 specific, parse-edid builds and
works fine on other architectures so make it available everywhere.

Also drop the custom install step as 'make install' does the right thing.
This does cause us to install into /usr/sbin instead of /sbin, but as that
is what upstream wants we can consider that a bugfix.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>