br2-external: make it available to post-{build, image} scripts
Currently, BR2_EXTERNAL is not always exported to sub-processes that we
spawn, like post-build or post-image scripts. This all depends on how
the user passes BR2_EXTERNAL; consider the following:

  - make BR2_EXTERNAL=/path/to/br2-ext
  - BR2_EXTERNAL=/path/to/br2-ext make

In the first case, it is just a make variable, not an environment
variable, and thus not exported, while in the second case it is an
environment variable and gets exported to all sub-processes make may
spawn.

Explicitly export it using EXTRA_ENV.

Reported-by: Julian Scheel <julian@jusst.de>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Julian Scheel <julian@jusst.de>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>