package/vlc: express the dependency on freerdp
vlc can use the freerdp library if present.

This is currently a hidden dependency, as we do not express it in our
vlc.mk rules. Package ordering makes it so that freerdp is built before
vlc; however, manually building vlc (with 'make vlc') would miss the
dependency because of that package ordering.

Explcitly add freerdp (if enabled) as a dependency of vlc to ensure
build order in every cases.

Reported-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>