mjpg-streamer: Fix build with ccache enabled.
Quote the TARGET_CC variable, otherwise, if ccache is enabled:

[...] make CC=$TARGET_CC [...]
is expanded to e.g.
[...] make CC=ccache gcc [...]

The build succeeds with "make[2]: Nothing to be done for '[...]gcc'".
and the program is compiled at "make install" time, with the default
host compiler, resulting in a binary that does not run on the target.

Signed-off-by: Nicolas Cavallari <nicolas.cavallari@green-communications.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>