radvd: remove stray comment
In commit 9863553fe8fa16a4d19f652b409a44bb58ef5567 ("packages: all
salute the passing of avr32"), the dependency on !BR2_avr32 was
removed, but not the comment above it which explained why radvd would
not build on AVR32. So now, we have the impression that the comment
applies to the "select BR2_PACKAGE_FLEX" line, which doesn't make
sense.

Therefore, this commit gets rid of the unneeded and confusing comment.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>